#936: add construct URL brick
